+++ app/models/campaign/channel.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class Campaign
+  class Channel < ApplicationRecord
+    belongs_to :campaign
+    belongs_to :channel
+  end
+end