+++ app/models/campaign/character.rb
@@ -2,7 +2,7 @@
 
 class Campaign
   class Character < ApplicationRecord
-    belongs_to :campaign
-    belongs_to :character
+    belongs_to :campaign, class_name: '::Campaign'
+    belongs_to :character, class_name: '::Character'
   end
 end