+++ app/services/bot_context/handle_command_service.rb
@@ -6,7 +6,8 @@ module BotContext
       roll_command: 'services.bot_context.commands.roll',
       homebrew_command: 'services.bot_context.commands.homebrew',
       book_command: 'services.bot_context.commands.book',
-      check_command: 'services.bot_context.commands.check'
+      check_command: 'services.bot_context.commands.check',
+      campaign_command: 'services.bot_context.commands.campaign'
     ]
 
     def call(source:, command:, arguments:, data:)
@@ -15,6 +16,7 @@ module BotContext
       when '/homebrew' then homebrew_command.call(source: source, arguments: arguments, data: data)
       when '/module' then book_command.call(source: source, arguments: arguments, data: data)
       when '/check' then check_command.call(source: source, arguments: arguments, data: data)
+      when '/campaign' then campaign_command.call(source: source, arguments: arguments, data: data)
       end
     end
   end